% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gMulti.R
\name{gMulti}
\alias{gMulti}
\title{Graph-based multiple change-point detection}
\usage{
gMulti(y, search.type = "WBS", model.selection = TRUE, dendrogram = TRUE, ...)
}
\arguments{
\item{y}{A n x d matrix to be scanned with n observations and d dimensions.}

\item{search.type}{Greedy method to be performed in the first step.\preformatted{ "WBS" specifies graph-based wild binary segmentation

 "SBS" specifies graph-based seeded binary segmentation
}}

\item{model.selection}{Perform the pruning (model selection) step.}

\item{dendrogram}{Plot change-point dendrogram}

\item{...}{Arguments to be passed to \code{\link[gMulti]{gWBS}}, \code{\link[gMulti]{gSBS}}, \code{\link[gMulti]{gBE}} and \code{\link[gMulti]{cpdendrogram}}}
}
\value{
\item{tauhat}{Final selected change-points after pruning.}
\item{tautilde}{Searched change-points in the first step.}
}
\description{
This function performs gMulti: multiple change-point detection by graph-based method. The whole process involves 1. searching and 2. pruning step. A change-point dedrogram will be plotted for understanding structure of detected change-points.
}
\examples{
set.seed(1)
y <- rbind(
  matrix(rnorm(200), ncol = 10), matrix(rexp(200), ncol = 10),
  matrix(rnorm(200), ncol = 10), matrix(rexp(200), ncol = 10)
)
gMulti(y)
}
